Refactor and rearrange the code
