+++ README.md
@@ -5,7 +5,7 @@
 You should have:
 - node.js installed (version 8+ or the one that supports ES6+)
 - Chrome or other browser that supports ES6 and Notification API
-- Gerrit, duh
+- Gerrit, with Jenkins integration (maybe message annotation would work with other CIs, though)
 
 ## How to run
 
@@ -14,6 +14,18 @@ Clone the project and rename `configuration.example.js` to `configuration.js` an
 `npm start`
 
 It will start an instance of express.js server on port 3000. You should be able to open browser on `http://localhost:3000/ui` to display UI. First you will be required to enable notifications in order to receive some.
+
+## Dev mode
+
+If you don't have Gerrit instance available, you can start in development mode. You need to run to separate development scripts that would serve the frontend and backend parts separately. 
+
+`npm start:dev`
+
+and
+
+`npm bundle:watch`
+
+This would start the webpack watcher that will build the changes as you save files. You need to refresh the page manually, though :).
 
 ## Contribution
