+++ ui/app.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Menu from './menu';
-import CommentList from './comment-list';
+import CommentList from './comments/comment-list';
 import {changeState, markAllAsRead} from "./actions";
 
 const app = (state) => (<div>
